requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.28.0
pillow>=9.0.0
requests>=2.28.0
numpy>=1.21.0